change the sematics of required and default
The implication of required is now equivalent to saying that
a given field must always be present when a model is sent
over the wire. Fields with defaults are required fields
for which generated code provides a default value.

All non-required fields are represented as options in Scala.

All non-required fields are represented as nullable in Ruby.

Non-required fields in Scala are now serialized using
writeNullable so that they are not mapped to null in
the JSON representation (https://github.com/gilt/apidoc-generator/issues/7).

Notably, non-required collections now map to Option in Scala
and nullable in Ruby. They no longer default to an empty instance
of the collection.

Default values are parsed and injected as language specific
literals client side.

It is illegal to specify a default for a non-required
field.

Fixes https://github.com/gilt/apidoc-generator/issues/5
Fixes https://github.com/gilt/apidoc-generator/issues/7