Remove match from anonymous function generation.
You can replace an anonymous function which consists only
of a match, with a partial function instead. The code is a
bit more succinct, and less prone to having linters complain.
In this case, we can also be sure that all possible values are
matched, because of the catch-all option.