Fix problem with String/Enumerable in Ruby client
The existing code doesn't work in Ruby >= 1.9.3. In earlier versions of Ruby,
String inherited from Enumerable and thus #each and #map were both available.
In later versions, this is not the case - #each is available on String but not
and will work in both old and new versions of Ruby.