Android union generator: setting JsonTypeInfo.Id.NONE on child of union
This is necessary for a rare situation where union's child class is included
in json, not the union parent.  For example:

-----------------------------
  "enums": {
    "animal": {
      "discriminator": "type",
      "types": [
        {
          "type": "dog"
        },
        {
          "type": "cat"
        }
      ]
    }
  }
    ...
  "models": {
      "dog-house": {
      "fields": [
        {
          "name": "owner",
          "type": "dog",
          "description": "Only dog can be an owner of a dog house, not just any animal",
        }
      }
  }
-----------------------------

Without JsonTypeInfo.Id.NONE anotation class Dog implements Animal, and so it inherits Animal's jackson annotations.  And Animal's jackson annotations are looking for discriminator "type", which is not there.

With JsonTypeInfo.Id.NONE annotation Dog overrides Animal's annotation.