issue-415: support http4s client form params
there are a number of ways to do this.  I have ported
the approach from the play generator.  will not
address consistency between params in the url
and body. play generator has the same problem.

http4s client test suite is inadequate.  can't
address that either at this time.  will need
someone to do do an adhoc test because I
do not have much of a dev env for the generators.