Go: fix deserialization of `[string]`
We can't call `ioutil.ReadAll()` on an `interface{}`. The code was only
previously used at the top level, when the provided reader is the entire
json body and we could just `ReadAll()` into a string.

Now this supports reading strings inside maps and lists, because we can
just convert the `interface{}` into a `string`.